backend/requirements.txt
# FastAPI 웹 프레임워크 및 서버
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6     # 파일 업로드 처리

# 데이터 검증 및 설정
pydantic==2.5.3
python-dotenv==1.0.0

# HTTP 클라이언트
httpx==0.27.0

# LLM 관련 라이브러리
openai==1.10.0

# 문서 처리
pymupdf==1.24.14  # PDF 텍스트 추출
pillow==10.3.0    # 이미지 처리

# 벡터 DB
qdrant-client==1.16.1

# Sparse embedding (BM25)
fastembed==0.4.2

# 리랭킹 (유사 판례 검색)
sentence-transformers==5.2.2

# 유틸리티
tqdm==4.66.1               # Progress bar for migration scripts
# 데이터베이스 (PostgreSQL with Supabase)
sqlalchemy==2.0.25         # ORM for database operations
psycopg2-binary==2.9.9     # PostgreSQL adapter for Python
supabase==2.10.0           # Supabase client for storage

# 인증 및 보안
passlib==1.7.4             # Password hashing utilities
bcrypt==4.0.1              # Bcrypt hashing (compatible with passlib 1.7.4)
python-jose[cryptography]==3.3.0  # JWT token creation and verification

# LLM 관련 라이브러리 (필요에 따라 주석 해제)
# openai==1.10.0
# anthropic==0.8.1
# langchain==0.1.0
# langchain-openai==0.0.2
